fix(node): rewrite the process.env polyfill
The previous polyfill was unnecessarily complicated and hard to understand.
Additionally it also used fallback on globalThis to read and write env variables, which
in the past was used to read env variables in Cloudflare Workers in the case where workers
were using the legacy service worker format. This style of workers is not actively used any
more so this fallback is not needed, and in fact is undesirable as we never want to
accidentaly write this globalThis.foo as a result of an assignment to process.env.foo.

The nitro utilized globalThis.__env__ fallback is preserved as to no break nitro users.