fix(node): properly polyfill Node.js `global` symbol
Since `global` comes mainly from Node.js, we should by default provide its polyfill
in the nodeless preset so that other presets don't need to define it.

For example Deno preset shouldn't need to define that it needs the global polyfill,
as we expect the target environment to compose `env(nodeless, deno)` together.