feat: support Preset#inject negative override / subtraction
Sometimes an environment might want to undo any kind of inject configuration provided by the previous
preset, because this environment guaratees that an API is provided natively and doesn't need to be
supported via injection.

An example is Cloudflare's workerd, which natively support Blob and global when the nodejs_compat flag
is turned on.